Rename TLBO_FINAL.m to TLBO_ANN.m
This is a MATLAB code that uses the Teaching Learning Based Optimization (TLBO) algorithm to optimize three hyperparameters of Artificial Neural Networks: the number of hidden layers, the size of each hidden layer, and the learning rate. Please note that this code requires an additional cost function, which is included in this repository under the file name "ann_cost.m".